Refactor document history and version views
