+++ c2corg_api/views/route.py
@@ -4,6 +4,7 @@ from c2corg_api.models import DBSession
 from c2corg_api.models.association import Association
 from c2corg_api.models.document import DocumentLocale, DocumentGeometry
 from c2corg_api.models.outing import schema_association_outing, Outing
+from c2corg_api.views.document_version import DocumentVersionRest
 from c2corg_api.views.outing import set_author
 from c2corg_api.models.utils import get_mid_point
 from cornice.resource import resource, view
@@ -156,7 +157,7 @@ class RouteRest(DocumentRest):
 
 
 @resource(path='/routes/{id}/{lang}/{version_id}', cors_policy=cors_policy)
-class RouteVersionRest(DocumentRest):
+class RouteVersionRest(DocumentVersionRest):
 
     @view(validators=[validate_id, validate_lang, validate_version_id])
     def get(self):
